debug.c: Don't include the functions if _DEBUG is disabled
When _DEBUG is disabled, the previous definitions of these functions in
debug.h conflict with the functions in this file.  To solve this, only
build the functions if _DEBUG is enabled.

Testing Done: all builds still work as well as they did before
Review URL: https://reviews.leaflabs.com/r/1005/
Reviewed by: li_jun, fparent

Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>