debug.h: use "real" functions when _DEBUG is disabled
We want to find the build warnings when we mess up the debug function
arguments even when _DEBUG is disabled, so use real functions for these,
not just #defines.  The end result is the same, but now we get
typechecking, which is essential.

Testing Done: all builds still work as well as they did before
Review URL: https://reviews.leaflabs.com/r/1006/
Reviewed by: fparent, li_jun

Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>