Re-org where to do data hashing
Modified the data_load API to do hash inside the call.
This change would not affect the SPIROM loading at all, except for
looking not as good as before. However, with UniPro data loading,
this would allow the load function to hash the data while waiting
for the next block, so it would improve the performance a lot!

Testing Done:  only SPI ROM loading was tested and it is not broken.
Bug Number:    SW-723
Review URL:    https://reviews.leaflabs.com/r/1037/
Reviewed by:   Eli Sennesh <esennesh@leaflabs.com>,
               Morgan Girling <girling_morgan@projectara.com>

Signed-off-by: Jun Li <li_jun@projectara.com>