Enable fake crypto function for simulation
Although crypto library is not integrated yet, we can
enable some fake code for simulation. The code checks
the name of the key used to sign the TFTF. If it finds
the key name starts with FAIL, it fails the signaturei
verification.

Also this change fixed some conditions used to verify
TFTF section when a signature section is included.

Testing Done:   The boot ROM (built with _SIMULATION=1)
                fails loading image when it is signed
                with a key name start with FAIL, and
                recognize an image signed with other key
                names as trusted image.
Bug Number:
Review URL:
Reviewed by:

Signed-off-by:   Jun Li <li_jun@projectara.com>