Bootrom: Eliminated writes to RO DME registers
Because the HAPS-62 and simulator do not pre-load the hardwired DME registers
and because the Ara_VID and Ara_PID are needed when verifying the tftf blobs,
we had previously fetched the values from e-Fuses and wrote the 2 DME
registers. This provided more "sanitary" tftf code. However, Toshiba queried
why we were writing to RO registers. We elected to instead cache the ara_vid
and ara_pid in the communication area and fetch them from there in the tftf
code.

NB. SNL & SNH have not been addressed by this change because we're awaiting
clarification from Toshiba. This will be addressed if needed in a future
change.

Testing Done: verified that previously-valid tftf's remained so
Bug Number: ENG-442
Review URL: https://reviews.leaflabs.com/r/1588/
Reviewed by: Eli Sennesh, Jun Li

Signed-off-by:  Morgan Girling <morgang@bsquare.com>