bootrom: fix test FB-22 (ignore unknown TFTF types?)
The TFTF header validation was overly enthusiastic about rejecting unknown
section types and was changed to ignore them instead. After further review,
it was decided to retain the original behavior and augment some of the
code documentation. The Bootrom (1st level FW) will be strict about the
types that it loads for the 2nd level FW. The 2nd level FW and beyond are
not in ROM and can be upgraded to handle any TFTF/FFFF extensions. Note that
the bootloader only verifies the types of the sections of the TFTF that it
is loading, and so will be oblivious of unknown types in TFTF blocks outside
of its purview.

Testing Done: N/A - documentation change.
Bug Number: trivial
Review URL:
Reviewed by: Marti Bolivar, Jun Li

Signed-off-by:  Morgan Girling <morgang@bsquare.com>