Bootrom: build tool: extend bin2verilog to generate .bytecount
Per a request from Toshiba, we now generate a .bitcount file in addition to
the .dat file. Because the .dat file contains an 8-byte signature derived from
the date, it was safest to create an extended .bin file containing the original
.bin contents padded out with FFFFFFFF up to the last 8 bytes, which contain
the serial number. From that, it creates the Verilog download file
(essentially a hex dump) and a .bitcount file, containing the count of "1" bits
in the extended .bin file. This is an ASCII decimal number.

Testing Done: Verified the generated files with hexdump
Bug Number: SW-1205
Review URL:
Reviewed by:

Signed-off-by:  Morgan Girling <morgang@bsquare.com>