Bootrom: Expose the DME logging as a build flag
Previously, DME logging to dbgserial was done on an ad hoc basis by enabling
the #define locally, but not checking it in. To better intergrate with the
drop tools, etc. this is now exposed through the "-dme" option in all of the
build tools.

Testing Done: Verified through make3sim that the _DME_LOGGING flag is passed
through, and that it is blocked when debug messages are turned off
Bug Number: SW-1205
Review URL: https://reviews.leaflabs.com/r/1836/
Reviewed by: Marti Bolivar, Jun Li

Signed-off-by:  Morgan Girling <morgang@bsquare.com>