Bootrom: Testing: Special s3fw image that verifies it loads
A special s3fw image that will be signed and flashed onto a HAPS for the
sole purpose of verifing that it ran and that the secret information is
available to it, though it goes to lengths to ensure that the secrets are
not released.

NOTES:
1. It must never be included in a normal build
2. It must never be released into the wild

Testing Done: Verified with a local signing key.
Bug Number: SW-1025
Review URL:
Reviewed by:

Signed-off-by:  Morgan Girling <morgang@bsquare.com>