Fix a linking issue
When the .text size is odd, there seems to be a problem with the
linker script, that would put the .text section at an odd address.
And since the instructions needs to be aligned, so extra padding
were added in front of the instructions, which makes .text grow into
the following section (.data).

To fix this issue, force the address of .text to be align to 4 bytes.

The root cause of why .text become odd sized is still not very clear.
I have done some experiment. The odd size came from the .rodata.
There are several questions un-answered:
1. Why some "const char abc[] = .....;" were linked into .rodata but
   the others were not.
2. For an odd size array in #1, why sometimes it is padded to 4 bytes
   but sometimes it is not.

However, even with the un-answered questions, looks like .text section
with odd size is a legitimate output. So we need to prepare for that
anyway.

And I was never able to reproduce this issue in ES2 boot ROM build due
to the un-answered questions, but theoretically it would have simular
issue. So I also made the change for ES2 boot ROM build.

NOTE: This is not a problem for either ES3 boot ROM or ES3 testing third
stage FW, since those two always have fixed address for the .text section

Testing Done:   Even with odd sized .text, it is placed at even numbered
                address which is safe.
Bug Number:     SW-1279
Review URL:     https://reviews.leaflabs.com/r/1890/
Reviewed by:    Morgan Girling <girling_morgan@projectara.com>
                Eli Sennesh <esennesh@leaflabs.com>

Signed-off-by:  Jun Li <li_jun@projectara.com>