Avoid using stack in resume code
After resume from standby, avoid using stack.
The original plan was that there would be more code in the resume
path in ROM, and some delays are much longer and counting nops is
not a good idea. But after discussing with Toshiba, some of the code
should run in workram, which then would be part of the later stage
FW instead of the boot ROM. So for this 0.1us delay left in ROM, we
could do counting nops.

Testing Done:   Well, the code still build. But I cannot really test
                it unless we send to Toshiba for simulation.
                And this code is for boot ROM update. So unless we
                are going for ES4, this change would never be used.
Bug Number:     SW-1544
Review URL:     https://reviews.leaflabs.com/r/1933/
Reviewed by:    Patrick Titiano, Fabien Parent

Signed-off-by:  Jun Li <li_jun@projectara.com>