Make it possible to specify an output dir
There is a variable called "OUTROOT" in the Makefile that could let think that
it is possible to specify the output dir for the compilation. As it turns out,
there are a few other locations in the files where the output dir is actually
hardcoded to be "build/".

This patch fixes these locations and allows a truly customizable output dir.
For example,

$ make OUTROOT=toto

now builds everything into the subfolder "toto/".

Testing Done: compiled
Bug Number: trivial
Review URL: https://reviews.leaflabs.com/r/2138/
Reviewed by: gregkh

Signed-off-by: Joel Porquet <joel@porquet.org>