Allow fake IMS in config data of S2FL
This is a debugging feature we want to add to S2FL
With this feature enabled, the developer could use a un-signed
second stage FW for debugging but still maintain the capability of
all the security feature enabled, and without the risk of leaking
valid secrets in any certain bridge chip.
Also, this gives a pre-defined IMS, instead of a random number read
from the chip. This would be helpful for debugging.

Testing Done:   When fake IMS is enabled, S2FL uses the faked IMS,
                whether it is signed or un-signed.
Bug Number:     SW-1547
Review URL:     https://reviews.leaflabs.com/r/2308/
Reviewed by:    Morgan Girling

Signed-off-by:  Jun Li <li_jun@projectara.com>