makefile: fix error when make clean in an unconfigured working copy
Because of the inclusion of several configuration-specific files in the
Makefile, it is impossible to do a make clean/distclean in an unconfigured
directory.

This patch adds some logic in order to detect when clean/distclean are called
and doesn't try to include the unexisting files.

As a side-effect, it make the FDK's life easier since it allows the clean rule
of the FDK to call make clean in an unconfigured bootrom directory without
getting an error.

Testing Done: compiled
Bug Number: trivial
Review URL: https://reviews.leaflabs.com/r/2204/
Reviewed by: esennesh, Morgan Girling, gregkh

Signed-off-by: Joel Porquet <joel@porquet.org>