Clarify the mailbox usage
In our original design, the use of mailbox is a way to enable E2EFC
between two parties. It is not intended to communicate which cport
is used for what purpose, especially if there are more than one cport
(other than the control cport) used.
We have decided not to use E2EFC, at least for bootrom. But because of
schedule reason, we did keep the mailbox communication code. Now we
need to seperate it from how the gbbot cport is determined.

Testing Done:   Build still works as before.
Bug Number:     Trivial
Review URL:     https://reviews.leaflabs.com/r/2457/
Reviewed by:    Eli Sennesh

Signed-off-by:  Jun Li <li_jun@projectara.com>