Fix collision mis-judgement
There is a small issue in the code that checkes if two sections in
TFTF have collision. The problem happens when two sections are side
by side and the section with lower address comes after the higher
one in TFTF.

This change fixes this problem. However, with the boot ROM in ES3,
this bug is always going to be there. So for the S2L code, add a
32-byte gap between .text and the config data to make sure this bug
can be avoided.

TBD: A better way to avoid this problem is modify create-tftf to
     sort the sections in TFTF based on their address.

Testing Done:   This bug was found when some S2L images I built not
                able to boot on ES3. So I debugged it on HAPS.
                1. With the change in TFTF parsing code, the image
                   that fails originally can work now.
                2. With the fix in the S2L linker script, all images
                   can run on ES3.
Bug Number:     Trivial
Review URL:     https://reviews.leaflabs.com/r/2704/
Reviewed by:    Morgan Girling

Signed-off-by:  Jun Li <li_jun@projectara.com>