configure: make it possible to configure from an external location
Right now, it's impossible to call ./configure without being in at the root of
the bootrom directory. This forces scripts to perform a "cd" before such as:
cd /path/to/bootrom && ./configure ...

This behavior is not always convenient for scripts, especially when
encapsulating commands into logging and debugging calls. This patch thus makes
it possible to call configure from anywhere, as it will determine its own
location and perform the proper actions accordingly.

The output of 'cd' is redirected to /dev/null to play nice with those who use
CDPATH.  Using CDPATH makes 'cd' display the path after changing to a new
directory. This causes an issue with the way SCRIPT_PATH is determined.

This patch is in preparation of the new FDK.

Testing Done: compiled
Bug Number: SW-2736
Review URL:
Reviewed by:

Signed-off-by: Joel Porquet <joel@porquet.org>